Add support for CSV files and Fashion MNIST data
Update RCN impl to support reading CSV files or performing a directory
walk. Current implementation is not very modular, and would be a good
idea in the future to update the `from` functions to support a generic
reader with some dedicated input format allowing the reading process to
be generalized.